+++ docs/modules.html
@@ -3,8 +3,24 @@
   <head>
     <meta charset="utf-8" />
     <meta http-equiv="x-ua-compatible" content="IE=edge" />
-    <title>byte.db</title>
-    <meta name="description" content="Documentation for byte.db" />
+    <!-- Primary Meta Tags -->
+<title>byte.db</title>
+<meta name="title" content="byte.db">
+<meta name="description" content="Documentation for byte.db">
+
+<!-- Open Graph / Facebook -->
+<meta property="og:type" content="website">
+<meta property="og:url" content="https://cloudteamdev.github.io/ByteDatabase/">
+<meta property="og:title" content="byte.db">
+<meta property="og:description" content="Documentation for byte.db">
+<meta property="og:image" content="">
+
+<!-- Twitter -->
+<meta property="twitter:card" content="summary_large_image">
+<meta property="twitter:url" content="https://cloudteamdev.github.io/ByteDatabase/">
+<meta property="twitter:title" content="byte.db">
+<meta property="twitter:description" content="Documentation for byte.db">
+<meta property="twitter:image" content="">
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="stylesheet" href="assets/style.css" />
     <link rel="stylesheet" href="assets/highlight.css" />
