+++ lib/functions/index.ts
@@ -1,0 +1,2 @@
+export * from "./findTables";
+export * from "./sanitize";
