+++ lib/typings/index.ts
@@ -1,0 +1,4 @@
+export * from "./interfaces/DBP";
+export * from "./types/Data";
+export * from "./types/ToRecord";
+export * from "./types/Events";
