+++ lib/typings/interfaces/DBP.ts
@@ -1,5 +1,5 @@
 import { Options } from 'better-sqlite3';
 
 export interface DefaultByteProperties extends Options {
-    sanitize: boolean
+    sanitize: boolean | Boolean
 }
