+++ package.json
@@ -6,7 +6,10 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "prettier": "prettier --write .",
-    "test": "tsc && node dist/__tests__/func.js"
+    "test": "node dist/__tests__/func.js",
+    "build": "tsc",
+    "prepublish": "npm run build",
+    "pretest": "npm run build"
   },
   "keywords": [
     "database",
@@ -49,7 +52,7 @@
       "url": "https://github.com/Rubenennj"
     }
   ],
-  "homepage": "https://github.com/cloudteamedev/ByteDatabase",
+  "homepage": "https://cloudteamdev.github.io/ByteDatabase/",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/cloudteamdev/ByteDatabase.git"
